Update
Library compatibility with Python2.7:

line 389: "self._i2c.writeByte(self.address, self.SCMD_MA_DRIVE + motorNum, driveValue)" requires integer values
line 381: "level = int(round((level + 1 - direction)/2))" passes a float into "driveValue"

+++ qwiic_scmd.py
@@ -378,7 +378,7 @@ class QwiicScmd(object):
         """
 
         # Convert value to a 7-bit int and match the indexing for uint8_t values as needed in Arduino library
-        level = round((level + 1 - direction)/2)
+        level = int(round((level + 1 - direction)/2))
 
         driveValue = 0 # use to build value to actually write to register
